Extract JSON operator helpers
